% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsft_pdt.R
\name{nonpara_boot}
\alias{nonpara_boot}
\title{Nonparametric bootstrap to estimate the variance}
\usage{
nonpara_boot(
  t.pfs,
  t.co,
  t.os,
  delta.os,
  cen.time,
  a,
  mat.pdt,
  delta.co,
  delta.pdt,
  include.pdt = TRUE,
  const = 1,
  boot.function,
  tau.lower = -1,
  tau.upper = 1,
  grid.length = 50,
  B = 100
)
}
\arguments{
\item{t.pfs}{the progression-free survival (PFS) time.}

\item{t.co}{the observed time from the start to the end of the crossover
period.}

\item{t.os}{the overall survival (OS) time.}

\item{delta.os}{the event indicator of the OS.}

\item{cen.time}{the censoring time. Under administrative censoring, the
censoring time is available for each subject.}

\item{a}{the initial binary treatment indicator encoded as 0 (control) or
1 (treatment).}

\item{mat.pdt}{the matrix of all prognosis factors that may affect the
patient to receive the PDTs, with the dimension as n*p, where n: the sample
size, p: dimension of the prognosis factors. Inclusion of the survival times
before the PDT period is recommended.}

\item{delta.co}{the vector of crossover indicators.}

\item{delta.pdt}{the vector of the PDT indicators.}

\item{include.pdt}{a logical variable to indicate the inclusion of the PDT
effects in the OS analysis. Default: \code{include.pdt = TRUE}.}

\item{const}{the multiplicative sensitivity parameter that reflects a change
in the treatment benefit for subjects in the control group with treatment
crossover.}

\item{boot.function}{the function that is used to estimate the treatment
effect. Available functions include \code{rpsft.cox} and \code{rpsft.ipcw}.}

\item{tau.lower}{the lower bound of the treatment effect parameter for the
grid search. Default: \code{tau.lower = -1}.}

\item{tau.upper}{the upper bound of the treatment effect parameter for the
grid search. Default: \code{tau.upper = 1}.}

\item{grid.length}{the number of values in the grid search. Default:
\code{grid.length = 50}.}

\item{B}{the number of bootstrap replicates.}
}
\value{
tau.est an estimate of the treatment effect
}
\description{
This function implements the proposed RPSFT + IPCW method and returns an
estimate of the treatment effect on the overall survival. It can be used if
investigators have sufficient knowledge of the confounding factors that may
affect the probability of receiving the PDTs.
}
