% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsft_pdt.R
\name{rpsft.ipcw}
\alias{rpsft.ipcw}
\title{Implement the RPSFT + IPCW method}
\usage{
rpsft.ipcw(
  t.pfs,
  t.co,
  t.os,
  delta.os,
  cen.time,
  a,
  mat.pdt,
  delta.co,
  delta.pdt,
  include.pdt = TRUE,
  const = 1,
  tau.lower = -1,
  tau.upper = 1,
  grid.length = 50,
  var.est = FALSE,
  B = NULL
)
}
\arguments{
\item{t.pfs}{the progression-free survival (PFS) time.}

\item{t.co}{the observed time from the start to the end of the crossover
period.}

\item{t.os}{the overall survival (OS) time.}

\item{delta.os}{the event indicator of the OS.}

\item{cen.time}{the censoring time. Under administrative censoring, the
censoring time is available for each subject.}

\item{a}{the initial binary treatment indicator encoded as 0 (control) or
1 (treatment).}

\item{mat.pdt}{the matrix of all prognosis factors that may affect the
patient to receive the PDTs, with the dimension as n*p, where n: the sample
size, p: dimension of the prognosis factors. Inclusion of the survival times
before the PDT period is recommended.}

\item{delta.co}{the event indicator of treatment crossover.}

\item{delta.pdt}{the event indicator of the PDTs.}

\item{include.pdt}{a logical variable to indicate the inclusion of the PDT
effects in the OS analysis. Default: \code{include.pdt = TRUE}.}

\item{const}{the multiplicative sensitivity parameter that reflects a change
in the treatment benefit for subjects in the control group with treatment
crossover.}

\item{tau.lower}{the lower bound of the treatment effect parameter for the
grid search. Default: \code{tau.lower = -1}.}

\item{tau.upper}{the upper bound of the treatment effect parameter for the
grid search. Default: \code{tau.upper = 1}.}

\item{grid.length}{the number of values in the grid search. Default:
\code{grid.length = 50}.}

\item{var.est}{the logical variable to indicate whether the bootstrap
variance estimate will be included. Default: \code{var.est = FALSE}, i.e., no
inclusion of the variance estimate.}

\item{B}{the number of bootstrap replicate. When setting \code{var.est = TRUE},
\code{B} must be given.}
}
\value{
a list of test results
\itemize{
\item tau.est: the estimate of the treatment effect
\item var.est: the variance estimate of the treatment effect if set
\code{var.est = TRUE}
}
}
\description{
This function implements the proposed RPSFT + IPCW method and returns an
estimate of the treatment effect on the overall survival. It can be used if
investigators have sufficient knowledge of the confounding factors that may
affect the probability of receiving the PDTs.
}
\examples{
x <- cbind(dat$x1, dat$x2)
a <- dat$a
t.pfs <- dat$t.pfs
t.co <- dat$t.co
t.os <- dat$t.os
cen.time <- dat$c
delta.pfs <- dat$delta.pfs
delta.dp <- dat$delta.dp
delta.pdt <- dat$delta.pdt
delta.co <- dat$delta.co
delta.os <- dat$delta.os
mat.pdt <- cbind(x, t.co)
res.rpsftipcw <- rpsft.ipcw(t.pfs, t.co, t.os, delta.os,
                          cen.time, a, mat.pdt, delta.co, delta.pdt,
                          include.pdt = TRUE, const = 1,
                          tau.lower = -3, tau.upper = 1,
                          grid.length = 100)
res.rpsftipcw
}
