% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsft_pdt.R
\name{rpsft.linear}
\alias{rpsft.linear}
\title{Fit the RPSFT model on the original data with the estimated PDT effects}
\usage{
rpsft.linear(
  t.pfs,
  t.co,
  t.os,
  a,
  delta.os,
  delta.co,
  cen.time,
  pdt.effect,
  const = 1,
  weights = rep(1, length(a)),
  tau.lower = -1,
  tau.upper = 1,
  grid.length = 50
)
}
\arguments{
\item{t.pfs}{the progression-free survival (PFS) time.}

\item{t.co}{the observed time from the start to the end of the crossover
period.}

\item{t.os}{the overall survival (OS) time.}

\item{a}{the initial binary treatment indicator encoded as 0 (control) or
1 (treatment).}

\item{delta.os}{the event indicator of the OS.}

\item{delta.co}{the event indicator of treatment crossover.}

\item{cen.time}{the censoring time. Under administrative censoring, the
censoring time is available for each subject.}

\item{pdt.effect}{a vector of the estimated PDT effects.}

\item{const}{the multiplicative sensitivity parameter that reflects a change
in the treatment benefit for subjects in the control group with treatment
crossover. Default: \code{const = 1}.}

\item{weights}{the use-defined weights used in the log-rank test. Default:
each individual has an equal weight, i.e., \code{weights = rep(1, length(a))}.}

\item{tau.lower}{the lower bound of the treatment effect parameter for the
grid search. Default: \code{tau.lower = -1}.}

\item{tau.upper}{the upper bound of the treatment effect parameter for the
grid search. Default: \code{tau.upper = 1}.}

\item{grid.length}{the number of values in the grid search. Default:
\code{grid.length = 50}.}
}
\value{
a list of test results
\itemize{
\item test.stat: the test statistic
\item p.value: the p-value
\item tau: the estimate that results in a zero test statistic
\item test.stat.seq: a sequence of test statistics in the grid search
\item p.value.seq: a sequence of p-values in the grid search
}
}
\description{
This function fits the RPSFT model and returns the treatment effect estimate
and the corresponding p-value which result in a zero test statistic of the
log-rank test. It is an intermediate function that corresponds to the second
step of the proposed RPSFT + Cox method to account for the treatment
crossover. It can be used if investigators have sufficient knowledge of the
PDT effects and the survival time during the PDT period.
}
